pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ovh_exporter"
description = "ovh_exporter tool"
readme = "README.md"
dynamic = ["version"]
requires-python = ">=3.8"
license = {text = "Apache-2.0"}
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Natural Language :: English",
  "Environment :: Console",
  "Topic :: Utilities",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12"
]
dependencies = [
  "click~=8.1.6",
  "ovh~=1.1",
  "PyYAML~=6.0",
  "prometheus_client<1",
  "python-dotenv~=1.0",
  "gunicorn"
]

[[project.authors]]
name = "Laurent Almeras"
email = "lalmeras@gmail.com"

[project.urls]
repository = "https://github.com/lalmeras/ovh_exporter"
changelog = "https://github.com/lalmeras/ovh_exporter/blob/main/README.md"

[project.scripts]
ovh_exporter = "ovh_exporter.cli:main"

[tool.hatch.version]
path = "src/ovh_exporter/__init__.py"


[tool.hatch.envs.pip-review]
type = "container"
python = "3.8"
dependencies = [
  "pip-review"
]

[tool.hatch.envs.pip-review.scripts]
dev-pip-review = "pip-review"


[tool.hatch.envs.dev]
dependencies = [
  "types-requests",
  "types-PyYAML",
  "mypy",
  "rope",
  "ruff",
  "ipdb",
  "deptry"
]


[tool.hatch.envs.test]
type = "container"
dependencies = [
  "pytest",
  "pytest-cov"
]

[[tool.hatch.envs.test.matrix]]
python = ["3.8", "3.12"]


[tool.pytest.ini_options]
minversion = "7.1"
addopts = "--cov-config .coveragerc --cov-report=xml --cov=."
testpaths = [
    "tests"
]


[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
files = "src/ovh_exporter/**/*.py"

[[tool.mypy.overrides]]
module = [
  "path_dict",
  "ovh"
]
ignore_missing_imports = true


[tool.rope]
split_imports = true
